添加rpc接口
