+++ pom.xml
@@ -38,12 +38,6 @@
 		    <version>2.8.5</version>
 		</dependency>
 		
-		<dependency>
-			<groupId>commons-codec</groupId>
-			<artifactId>commons-codec</artifactId>
-			<version>1.9</version>
-		</dependency>
-		
 	</dependencies>
 
 	<build>
