+++ src/main/java/qlc/mng/P2PMng.java
@@ -8,16 +8,4 @@ import qlc.network.QlcException;
 import java.io.IOException;
 
 public class P2PMng {
-    /**
-     * Return online representative accounts that have voted recently
-     * @param url
-     * @param params null
-     * @return
-     * @throws QlcException
-     * @throws IOException
-     */
-    public static JSONObject onlineRepresentatives(String url, JSONArray params) throws QlcException, IOException {
-        QlcClient client = new QlcClient(url);
-        return client.call("net_onlineRepresentatives", params);
-    }
 }
