+++ src/test/java/test/AccountTest.java
@@ -2,8 +2,8 @@ package test;
 
 import java.util.UUID;
 
-import org.junit.Test;
 
+import org.junit.Test;
 import qlc.mng.AccountMng;
 import qlc.utils.Helper;
